FindSteadyStates

FindSteadyStates.Log_uniform

FindSteadyStates.ParameterGrid

FindSteadyStates.Uniform

Base.getindex

FindSteadyStates.get_sol2array

FindSteadyStates.rand_vec

FindSteadyStates.rand_vec

FindSteadyStates.rand_vec

FindSteadyStates.rand_vecU

FindSteadyStates.rand_vecU

FindSteadyStates.solve_SSODE_threads

FindSteadyStates.Log_uniform — Type

Sampler for log uniform distribution. a,b ∈ (0,∞)

FindSteadyStates.ParameterGrid — Type

Grid search iterator for parameters. The sequence of ranges defines the grids to search.

Parameters

ParameterGrid: Contain list of ranges. The range is in (start, end, points) order.

Examples

julia> ranges = [ (1.,10.,10.), (1.,10.,10.) ] # list of ranges (start_num, stop_num, number of grids`{int}`)
julia> param_range = ParameterRange(ranges)

FindSteadyStates.Uniform — Type

Sampler for uniform distribution

Base.getindex — Method

Get the parameters that would be ind th in iteration

The total sample number is # to do

Paremeters

ind{int}: The iteration index

Returns

params: list of numbers in the range of ParameterGrid.param_ranges

FindSteadyStates.get_sol2array — Method

Get the vector of vectors of results. sol can be :EnsembleSolution or other solutions from DifferentialEquations.solve

FindSteadyStates.rand_vec — Method

General method for sampling a vector with individual samplers

FindSteadyStates.rand_vec — Method

generate random vector with uniform distribution rand()

FindSteadyStates.rand_vec — Method

Sampling a vector with a list of sampler

REPL

unis = [Uniform(1,2) for i in 1:10]
vec = rand_vec(unis)

FindSteadyStates.rand_vecU — Method

General method. domains can be list of tuples

FindSteadyStates.rand_vecU — Method

Uniform sampling of vector

FindSteadyStates.solve_SSODE_threads — Method

Multi-thread version of steady-state solver

Arguements

func: DE function

us: Vector of vectors of initial variables

p: parameter constant