FindSteadyStates

FindSteadyStates.DEsteady

FindSteadyStates.ParameterGrid

FindSteadyStates.ParameterRandom

FindSteadyStates.jacobian

FindSteadyStates.ParameterGrid — Type

Grid search iterator for parameters. The sequence of ranges defines the grids to search.

Parameters

ParameterGrid: Contain list of ranges. The range is in (start, end, points) order.

Examples

julia> ranges = [ (1.,10.,10.), (1.,10.,10.) ] # list of ranges (start_num, stop_num, number of grids`{int}`)
julia> param_range = ParameterRange(ranges)

julia> ParameterGrid([ [1,1000,5], [1,3,1]]; method=LogGrid())
5-element ParameterGrid:
 [1.0999, 3.0]
 [1.999, 3.0]
 [10.99, 3.0]
 [100.9, 3.0]
 [1000.0, 3.0]

FindSteadyStates.ParameterRandom — Type

Set all the variables with same method with defined ranges

Parameter Random sampling with range.

Reset one specified method

FindSteadyStates.DEsteady — Type

Struct for solving steady state of an differential equation model.

Argument

func: ODE function.

u0: initial values

p: parameters

method: Method for solving steady-states. (i.e. DifferentialEqiations.Tsit5(), DifferentialEquations.AutoTsit5(Rosenbrock23()))

Reference

ODE solvers of DifferentialEquations.jl

FindSteadyStates.jacobian — Type

Construct the jacobian struct type

Argument

ode {ODE function}: with the form f(du,u,p,t)

u {Array}: initial values

p {Array}: parameters

Return

Jacobian {struct}

Warning

NameTuple definition of ode function is unrecommended. Due to the incompatibility with ModelingToolkit.modelingtoolkitize

Use DEmeta for jacobian generation

Calculate the jacobian of given initial valuables and parameters.

Get jacobian from state variable with default paramter set

Get jacobian from default state.