FindSteadyStates

FindSteadyStates.DEsteady

FindSteadyStates.DEsteady

FindSteadyStates.Log_uniform

FindSteadyStates.ODEtime

FindSteadyStates.ODEtime

FindSteadyStates.ParameterGrid

FindSteadyStates.ParameterGrid

FindSteadyStates.ParameterGrid

FindSteadyStates.Uniform

FindSteadyStates.UniformGrid

Base.getindex

DiffEqBase.solve

FindSteadyStates.get_sol2array

FindSteadyStates.mul

FindSteadyStates.mul

FindSteadyStates.rand_vec

FindSteadyStates.rand_vec

FindSteadyStates.rand_vec

FindSteadyStates.rand_vecU

FindSteadyStates.rand_vecU

FindSteadyStates.recursive_index

FindSteadyStates.update_struct_broadcast

FindSteadyStates.DEsteady — Type

Struct for solving steady state of an differential equation model.

Argument

func: ODE function.

u0: initial values

p: parameters

method: Method for solving steady-states. (i.e. DifferentialEqiations.Tsit5(), DifferentialEquations.AutoTsit5(Rosenbrock23()))

Reference

ODE solvers of DifferentialEquations.jl

FindSteadyStates.DEsteady — Method

Update steady state meta, and return another DEsteady object.

julia> using LabelledArrays, DifferentialEqiations
julia> deS = DEsteady(func=x->x, u0= LVector(s1=1.0,s2=2.0), p=1.0)
DEsteady
  func: #17 (function of type var"#17#18")
  p: Float64 1.0
  u0: LArray{Float64,1,Array{Float64,1},(:s1, :s2)}
  method: Tsit5 Tsit5()

julia> deS_new = deS([1000.0,200.0];key=:u0)
DEsteady
  func: #17 (function of type var"#17#18")
  p: Float64 1.0
  u0: LArray{Float64,1,Array{Float64,1},(:s1, :s2)}
  method: Tsit5 Tsit5()

julia> deS_new.u0
2-element LArray{Float64,1,Array{Float64,1},(:s1, :s2)}:
 :s1 => 1000.0
 :s2 => 200.0

FindSteadyStates.Log_uniform — Type

Sampler for log uniform distribution. a,b ∈ (0,∞)

FindSteadyStates.ODEtime — Type

Struct for solveing time-series of differential-equations. The data type of ode function is referenced to DifferentialEquations.jl.

Argument

func: ODE function.

u0: initial values

p: parameters

tspan: time span

References

DifferentialEquations.jl

FindSteadyStates.ODEtime — Method

Reset a field of ODEtime struct with broadcast method.

Usage

When using the LabelledArray.jl to define function and subtypes of DEmeta. Use this method to update the values of named arrays without changing the type of the field.

Purpose

This feature is to solve the LabelledArray problem. when using the field vector to define function, the initial values or parameters need to be Named vectors. However, the grid search iterator returns vector which gets error when apply with the funtion of name vector.

Arguement

u0: should be vector or types that can be broadcast. The length of u0 should be same as ODEtime.u0

key: field name of the stuct (default: :u0).

Example

julia> using LabelledArrays, DifferentialEqiations
julia> u = LVector(s1=1.0,s2=0.2)
2-element LArray{Float64,1,Array{Float64,1},(:s1, :s2)}:
 :s1 => 1.0
 :s2 => 0.2

julia> de = ODEtime(func=x->x, u0=u, p=1.0, tspan=(0.0,1.0))
ODEtime
  func: #9 (function of type var"#9#10")
  u0: LArray{Float64,1,Array{Float64,1},(:s1, :s2)}
  p: Float64 1.0
  tspan: Tuple{Float64,Float64}
  method: CompositeAlgorithm{Tuple{Tsit5,Rosenbrock23{0,true,DefaultLinSolve,DataType}},AutoSwitch{Tsit5,Rosenbrock23{0,true,DefaultLinSolve,DataType},Rational{Int64},Int64}}

julia> de_new = de([1.3,1.4];key=:u0)
ODEtime
  func: #9 (function of type var"#9#10")
  u0: LArray{Float64,1,Array{Float64,1},(:s1, :s2)}
  p: Float64 1.0
  tspan: Tuple{Float64,Float64}
  method: CompositeAlgorithm{Tuple{Tsit5,Rosenbrock23{0,true,DefaultLinSolve,DataType}},AutoSwitch{Tsit5,Rosenbrock23{0,true,DefaultLinSolve,DataType},Rational{Int64},Int64}}

julia> de_new.u0 # The updated u0 is the struct of LArray
2-element LArray{Float64,1,Array{Float64,1},(:s1, :s2)}:
 :s1 => 1.3
 :s2 => 1.4

julia> de.u0
2-element LArray{Float64,1,Array{Float64,1},(:s1, :s2)}:
 :s1 => 1.0
 :s2 => 0.2

FindSteadyStates.ParameterGrid — Type

Grid search iterator for parameters. The sequence of ranges defines the grids to search.

Parameters

ParameterGrid: Contain list of ranges. The range is in (start, end, points) order.

Examples

julia> ranges = [ (1.,10.,10.), (1.,10.,10.) ] # list of ranges (start_num, stop_num, number of grids`{int}`)
julia> param_range = ParameterRange(ranges)

FindSteadyStates.ParameterGrid — Method

Grid object with ranges and distribution function

FindSteadyStates.ParameterGrid — Method

Construct parameter grid from list of ranges ([start, end, grid_num]). The grid distribution is default to be uniform.

FindSteadyStates.Uniform — Type

Sampler for uniform distribution

FindSteadyStates.UniformGrid — Type

get the number of a given range from index with uniform distribution.

julia> FindSteadyStates.uniformGrid(1,10,3, 2)
5

Base.getindex — Method

Get the parameters that would be ind th in iteration

The total sample number is # to do

Paremeters

ind {int}: The iteration index

Returns

params: list of numbers in the range of ParameterGrid.param_ranges

DiffEqBase.solve — Method

Multi-thread version of steady-state solver

Arguements

func: DE function

us: Vector of vectors of initial variables

p: parameter constant

FindSteadyStates.get_sol2array — Method

Get the vector of vectors of results. sol can be :EnsembleSolution or other solutions from DifferentialEquations.solve

FindSteadyStates.mul — Method

Multipication of series with type converting.

FindSteadyStates.mul — Method

Multiplication of series.

FindSteadyStates.rand_vec — Method

General method for sampling a vector with individual samplers

FindSteadyStates.rand_vec — Method

generate random vector with uniform distribution rand()

FindSteadyStates.rand_vec — Method

Sampling a vector with a list of sampler

REPL

unis = [Uniform(1,2) for i in 1:10]
vec = rand_vec(unis)

FindSteadyStates.rand_vecU — Method

General method. domains can be list of tuples

FindSteadyStates.rand_vecU — Method

Uniform sampling of vector

FindSteadyStates.recursive_index — Method

Recursive indexing. This function uses list of ranges and index to provide a vector of indexes.

Example

julia> ranges = [(1,2,2), (4,45,2)]
julia> param_range = ParameterGrid(ranges)
julia> vec_i = recursive_index(param_range, 4)
julia> vec_i
[2,2]

FindSteadyStates.update_struct_broadcast — Method

Update a field of struct varibale by broadcasting.

Argument

var_struc: A struct variable.

vals: A Vector which has the same length with Base.getfield(var_struc, key) for update the new struct variable.

key : A Symbol which can apply Base.getfield(var_struc, key)

Return

Struct variable similar to var_struc but with renew vals on key site.